requirements.txt
psutil
kafka-python
pyspark
redis
streamlit
pandas
scikit-learn
plyer
